✨ feat: 주문 반려 기능 구현

+++ src/main/java/com/x1/frans/order/command/domain/repository/StoreOrderDeadlineRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StoreOrderDeadlineRepository extends JpaRepository<StoreOrderDeadline, Integer> {
+public interface StoreOrderDeadlineRepository extends JpaRepository<StoreOrderDeadline, Long> {
 }
